Clean up tophat's src/Makefile.am
I'm forwarding this patch from PR #8244 for the Spack package manager,
https://github.com/spack/spack/pull/8244.

Tophat was failing to build with automake@1.16.1, it worked with the
older automake@1.15.1.  This commit adds a patch to Tophat's
`src/Makefile.am` which cleans up a few things.  The result builds
successfully with both automake@1.15.1 and automake@1.16.1.  I have no
way to check that the resulting builds Do The Right Thing.

It changes two things:

- the original Makefile.am had a rule for `$(SAMPROG)` with no
  actions, and since there was a directory there with a tempting name
  the newer automake was trying to do something with it, but lacked
  the appropriate clues.  Since that target is actually made as a side
  effect of making the library (sigh...), it seems to work to just
  give that rule something harmless to do (the Peter Principle
  triumphs again...).

- a bunch of the targets need a `libtophat.a` and `libgc.a`; the older
  automake was probably able to guess what to do given the list of
  sources but the newer automake apparently won't make the necessary
  assumptions.  This patch wires up a simple rule and cleans up the
  appropriate dependencies so that things work.

While it may appear that I'm someone who understands automake, keep in
mind that I only play such a person on a TV reality show.  YMMV.